solve error

+++ utils_flow/visualization_utils.py
@@ -151,7 +151,7 @@ def overlay_semantic_mask(im, ann, alpha=0.5, mask=None, colors=None, color=[255
     example usage:
     image_overlaid = overlay_semantic_mask(im.astype(np.uint8), 255 - mask.astype(np.uint8) * 255, color=[255, 102, 51])
     """
-    im, ann = np.asarray(im, dtype=np.uint8), np.asarray(ann, dtype=np.int)
+    im, ann = np.asarray(im, dtype=np.uint8), np.asarray(ann, dtype=np.int32)
     if im.shape[:-1] != ann.shape:
         raise ValueError('First two dimensions of `im` and `ann` must match')
     if im.shape[-1] != 3:
